requirements.txt
bcrypt==5.0.0
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
halo==0.0.31
invoke==2.2.1
Jinja2==3.1.6
log-symbols==0.0.14
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
netmiko==4.6.0
nornir==3.5.0
nornir-jinja2==0.2.0
nornir-netmiko==1.0.1
nornir-utils==0.2.0
ntc_templates==8.1.0
paramiko==4.0.0
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.2
pyserial==3.5
PyYAML==6.0.3
rich==14.2.0
ruamel.yaml==0.19.1
scp==0.15.0
six==1.17.0
spinners==0.0.24
termcolor==3.3.0
textfsm==2.1.0
tqdm==4.67.1
typing_extensions==4.15.0
